utils/containers: add concurrencykit MPMC FIFO wrapper
