+++ utils/containers/lock_free/tests/lf_mpmc_queue.c
@@ -1,0 +1,42 @@
+/*
+ * Copyright (c) 2018 IOTA Stiftung
+ * https://github.com/iotaledger/entangled
+ *
+ * Refer to the LICENSE file for licensing information
+ */
+
+#include <unity/unity.h>
+
+#include "utils/containers/lock_free/lf_mpmc_queue_int.h"
+
+void queue_integers_same_thread() {
+  lf_mpmc_queue_int_t queue;
+
+  TEST_ASSERT(lf_mpmc_queue_int_init(&queue) == RC_OK);
+
+  TEST_ASSERT_TRUE(LF_MPMC_QUEUE_IS_EMPTY(&queue));
+
+  for (int i = 0; i < 1000; ++i) {
+    TEST_ASSERT(lf_mpmc_queue_int_enqueue(&queue, &i) == RC_OK);
+  }
+
+  TEST_ASSERT_FALSE(LF_MPMC_QUEUE_IS_EMPTY(&queue));
+
+  int dequeued = 0;
+  for (int i = 0; i < 1000; ++i) {
+    TEST_ASSERT(lf_mpmc_queue_int_dequeue(&queue, &dequeued) == RC_OK);
+    TEST_ASSERT_EQUAL_INT(i, dequeued);
+  }
+
+  TEST_ASSERT_TRUE(LF_MPMC_QUEUE_IS_EMPTY(&queue));
+
+  TEST_ASSERT(lf_mpmc_queue_int_destroy(&queue) == RC_OK);
+}
+
+int main(void) {
+  UNITY_BEGIN();
+
+  RUN_TEST(queue_integers_same_thread);
+
+  return UNITY_END();
+}
